Separate Keyring interface and implementation
This makes it clearer what librepo.repo.Repository actually needs:
just a Mapping of role names to some private key representation
(maybe it could even be a plain Signer, but right now it's a bit
more).

The implementation should still be divided into two separate Keyrings
(EnvVarKeyring and InsecureFileKeyring, or something like that) because
they are never used together.
